+++ interops/c/libUWAPI/include/UWAPI_helpers.h
@@ -1,7 +1,7 @@
 #ifndef UWAPI_HELPERS_H
 #define UWAPI_HELPERS_H
 
-#include <UWAPI_interface.h>
+#include <UWAPI_plugin.h>
 
 void UWAPI_free_nonnull_position_stats(UWAPI_PositionStats *position_stats);
