+++ interops/c/plugin/example/.gitignore
